feat: auth improvement

+++ apps/shortify/src/app/shortify/shortify.service.ts
@@ -1,24 +1,32 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable, UnauthorizedException } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { nanoid } from 'nanoid';
+import { HttpService } from '@nestjs/axios'
 import { CreateShortifyDto } from './dto/create-shortify.dto';
 import { UpdateShortifyDto } from './dto/update-shortify.dto';
 import { Shortify } from './entities/shortify.entity';
+import { firstValueFrom } from 'rxjs';
 
 @Injectable()
 export class ShortifyService {
   private readonly domain: string = `${process.env.DOMAIN}:${process.env.PORT}/api/shortify/`;
 
   constructor(
-    @InjectRepository(Shortify) private readonly shortifyRepository: Repository<Shortify>,
+    @InjectRepository(Shortify) 
+    private readonly shortifyRepository: Repository<Shortify>,
+    private readonly httpService: HttpService,
   ) {}
+  
   async create(createShortifyDto: CreateShortifyDto, auth?: string) {
     const shortify: Shortify = new Shortify();
     shortify.url = createShortifyDto.url;
     shortify.shortId = await this.generateUniqueId()
     if (auth) {
+        const data = await this.getUserData(auth);
+        const { id } = data;
 
+        shortify.user = id;
     }
     this.shortifyRepository.save(shortify)
     return this.concatenate(shortify.shortId);
@@ -27,23 +35,55 @@ export class ShortifyService {
     const shortId = await this.generateUniqueId();
     return this.concatenate(shortId);
   }
-  findAll() {
-    return this.shortifyRepository.find();
-  }
+  async getUserData(auth: string) {
+    try {
+      const response = await firstValueFrom(
+        this.httpService.get(`${process.env.DOMAIN}:3001/auth`, {
+          headers: { Authorization: auth },
+        })
+      );
+      return response.data;
 
+    } catch (error) {
+      throw new UnauthorizedException('Não autorizado');
+    }
+  }
   findOneByShortId(shortId: string) {
     return this.shortifyRepository.findOneBy({ shortId });
   }
-
-  update(id: number, updateShortifyDto: UpdateShortifyDto) {
-    const shortify: Shortify = new Shortify();
-    shortify.url = updateShortifyDto.url;
-    shortify.id = id;
-    this.shortifyRepository.save(shortify)
+  async findAll(auth: string) {
+    const data = await this.getUserData(auth);
+    return this.shortifyRepository.find({
+      where: { user: data.id }, 
+    });
   }
 
-  remove(id: number) {
-    return this.shortifyRepository.delete(id);;
+
+  async update(id: number, updateShortifyDto: UpdateShortifyDto, auth: string) {
+    const data = await this.getUserData(auth);
+  
+    const shortify = await this.shortifyRepository.findOne({ where: { id, user: data.id } });
+  
+    if (!shortify) {
+      throw new ForbiddenException('Você não tem permissão para atualizar este item.');
+    }
+  
+    shortify.url = updateShortifyDto.url;
+    await this.shortifyRepository.save(shortify);
+  
+    return shortify;
+  }
+  
+
+  async remove(id: number, auth: string) {
+    const data = await this.getUserData(auth);
+  
+    const shortify = await this.shortifyRepository.findOne({ where: { id, user: data.id } });
+  
+    if (!shortify) {
+      throw new ForbiddenException('Você não tem permissão para remover este item.');
+    }
+    return this.shortifyRepository.delete(id);
   }
   async addClickCount(id: number): Promise<Shortify | null> {
     const shortify = await this.shortifyRepository.findOneBy({ id });
